fix: LLD refacto step alert box
